Add files via upload
This is self-project is still in the works. I am most likely going to redo this project and put it into an Android app because it will allow for my functionality goals. Its aim is to be a Restaurant Receipt that calculates the order of an individual with them interacting with a UI. The receipt contains nutritional facts and the price for the customer. For the restaurant, it's supposed to generate a QR that it can scan and upload to the register and order system.